+++ iap/iap.go
@@ -15,6 +15,7 @@ import (
 	"bytes"
 	"context"
 	"crypto"
+	"crypto/ecdsa"
 	"crypto/hmac"
 	"crypto/rsa"
 	"crypto/sha256"
@@ -25,6 +26,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"math/big"
 	"net/http"
 	"net/url"
 	"sort"
@@ -35,6 +37,24 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/heroiclabs/nakama-common/runtime"
 )
+
+const AppleRootPEM = `
+-----BEGIN CERTIFICATE-----
+MIICQzCCAcmgAwIBAgIILcX8iNLFS5UwCgYIKoZIzj0EAwMwZzEbMBkGA1UEAwwS
+QXBwbGUgUm9vdCBDQSAtIEczMSYwJAYDVQQLDB1BcHBsZSBDZXJ0aWZpY2F0aW9u
+IEF1dGhvcml0eTETMBEGA1UECgwKQXBwbGUgSW5jLjELMAkGA1UEBhMCVVMwHhcN
+MTQwNDMwMTgxOTA2WhcNMzkwNDMwMTgxOTA2WjBnMRswGQYDVQQDDBJBcHBsZSBS
+b290IENBIC0gRzMxJjAkBgNVBAsMHUFwcGxlIENlcnRpZmljYXRpb24gQXV0aG9y
+aXR5MRMwEQYDVQQKDApBcHBsZSBJbmMuMQswCQYDVQQGEwJVUzB2MBAGByqGSM49
+AgEGBSuBBAAiA2IABJjpLz1AcqTtkyJygRMc3RCV8cWjTnHcFBbZDuWmBSp3ZHtf
+TjjTuxxEtX/1H7YyYl3J6YRbTzBPEVoA/VhYDKX1DyxNB0cTddqXl5dvMVztK517
+IDvYuVTZXpmkOlEKMaNCMEAwHQYDVR0OBBYEFLuw3qFYM4iapIqZ3r6966/ayySr
+MA8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgEGMAoGCCqGSM49BAMDA2gA
+MGUCMQCD6cHEFl4aXTQY2e3v9GwOAEZLuN+yRhHFD/3meoyhpmvOwgPUnPWTxnS4
+at+qIxUCMG1mihDK1A3UT82NQz60imOlM27jbdoXt2QfyFMm+YhidDkLF1vLUagM
+6BgD56KyKA==
+-----END CERTIFICATE-----
+`
 
 const (
 	AppleReceiptValidationUrlSandbox    = "https://sandbox.itunes.apple.com/verifyReceipt"
@@ -149,8 +169,8 @@ type ValidateReceiptAppleResponse struct {
 }
 
 // Validate an IAP receipt with Apple. This function will check against both the production and sandbox Apple URLs.
-func ValidateReceiptApple(ctx context.Context, httpc *http.Client, receipt, password string) (*ValidateReceiptAppleResponse, []byte, error) {
-	resp, raw, err := ValidateReceiptAppleWithUrl(ctx, httpc, AppleReceiptValidationUrlProduction, receipt, password)
+func ValidateLegacyReceiptApple(ctx context.Context, httpc *http.Client, password, receipt string) (*ValidateReceiptAppleResponse, []byte, error) {
+	resp, raw, err := ValidateLegacyReceiptAppleWithUrl(ctx, httpc, AppleReceiptValidationUrlProduction, receipt, password)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -158,14 +178,14 @@ func ValidateReceiptApple(ctx context.Context, httpc *http.Client, receipt, pass
 	switch resp.Status {
 	case AppleReceiptIsFromTestSandbox:
 		// Receipt should be checked with the Apple sandbox.
-		return ValidateReceiptAppleWithUrl(ctx, httpc, AppleReceiptValidationUrlSandbox, receipt, password)
+		return ValidateLegacyReceiptAppleWithUrl(ctx, httpc, AppleReceiptValidationUrlSandbox, receipt, password)
 	}
 
 	return resp, raw, nil
 }
 
 // Validate an IAP receipt with Apple against the specified URL.
-func ValidateReceiptAppleWithUrl(ctx context.Context, httpc *http.Client, url, receipt, password string) (*ValidateReceiptAppleResponse, []byte, error) {
+func ValidateLegacyReceiptAppleWithUrl(ctx context.Context, httpc *http.Client, url, receipt, password string) (*ValidateReceiptAppleResponse, []byte, error) {
 	if len(url) < 1 {
 		return nil, nil, errors.New("'url' must not be empty")
 	}
@@ -897,4 +917,93 @@ func ValidateReceiptFacebookInstant(appSecret, signedRequest string) (*FacebookI
 	}
 
 	return payment, string(payload), nil
+}
+
+func ValidateAppleJwsSignature(receipt string) error {
+	jwsTokens := strings.Split(receipt, ".")
+	header := jwsTokens[0]
+	payload := jwsTokens[1]
+	signature := jwsTokens[2]
+
+	headerByte, err := base64.RawStdEncoding.DecodeString(header)
+	if err != nil {
+		return err
+	}
+
+	type Header struct {
+		Alg string   `json:"alg"`
+		X5c []string `json:"x5c"`
+	}
+	var jwsHeader Header
+	if err = json.Unmarshal(headerByte, &jwsHeader); err != nil {
+		return err
+	}
+
+	certs := make([][]byte, 0)
+	for _, encodedCert := range jwsHeader.X5c {
+		cert, err := base64.StdEncoding.DecodeString(encodedCert)
+		if err != nil {
+			return err
+		}
+		certs = append(certs, cert)
+	}
+
+	rootCert := x509.NewCertPool()
+	ok := rootCert.AppendCertsFromPEM([]byte(AppleRootPEM))
+	if !ok {
+		return err
+	}
+
+	leafCert, err := x509.ParseCertificate(certs[0])
+	if err != nil {
+		return err
+	}
+
+	interCert, err := x509.ParseCertificate(certs[1])
+	if err != nil {
+		return err
+	}
+	intermediates := x509.NewCertPool()
+	intermediates.AddCert(interCert)
+
+	cert, err := x509.ParseCertificate(certs[2])
+	if err != nil {
+		return err
+	}
+
+	opts := x509.VerifyOptions{
+		Roots:         rootCert,
+		Intermediates: intermediates,
+	}
+
+	_, err = cert.Verify(opts)
+	if err != nil {
+		return err
+	}
+
+	leafCertRsaPubKey, ok := leafCert.PublicKey.(*ecdsa.PublicKey)
+	if !ok {
+		return fmt.Errorf("failed to parse leaf certificate public key")
+	}
+
+	signingInput := []byte(fmt.Sprintf("%s.%s", header, payload))
+	hash := sha256.Sum256(signingInput)
+
+	sigBytes, err := base64.RawURLEncoding.DecodeString(signature)
+	if err != nil {
+		return fmt.Errorf("failed to decode jws signature: %w", err)
+	}
+
+	if len(sigBytes) != 64 {
+		return fmt.Errorf("invalid jws signature: invalid length %d, expected 64", len(sigBytes))
+	}
+
+	r := new(big.Int).SetBytes(sigBytes[:32])
+	s := new(big.Int).SetBytes(sigBytes[32:])
+
+	if !ecdsa.Verify(leafCertRsaPubKey, hash[:], r, s) {
+		return errors.New("invalid jws signature: ecdsa verification failed")
+	}
+
+	return nil
 }